Voting (#11)
Implemented a scheme of events to allow a class to send messages back to the chat room. This is currently in the voting script but should be moved out.

Created Vote class to handle logic of voting on polls via Twitch chat. A vote is started with !vote and answers are submitted with !answer